# openapi.yaml
openapi: 3.1.0
info:
    title: 'Marketplace API'
    version: 1.0.0
    description: 'Multi-vendor marketplace API — Express, TypeScript, Prisma. Auto-generated from Zod schemas.'
servers:
    -
        url: 'http://localhost:4000'
        description: 'Local development'
    -
        url: 'https://api.yourdomain.com'
        description: Production
security:
    -
        bearerAuth: []
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: 'Paste an **access token** here.'
...
paths:
...
    /customer/orders:
        get:
            tags:
                - customer
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    schema:
                        type: string
                        description: 'Search query'
                    required: false
                    description: 'Search query'
                    name: q
                    in: query
                -
                    schema:
                        type: string
                        pattern: ^\d+$
                        default: '1'
                        description: '1-based page number'
                    required: false
                    description: '1-based page number'
                    name: page
                    in: query
                -
                    schema:
                        type: string
                        pattern: ^\d+$
                        default: '20'
                        description: 'Number of items per page'
                    required: false
                    description: 'Number of items per page'
                    name: limit
                    in: query
                -
                    schema:
                        type: string
                        enum:
                            - PENDING
                            - PROCESSING
                            - SHIPPED
                            - DELIVERED
                            - CANCELLED
                        description: 'Filter orders by status'
                    required: false
                    description: 'Filter orders by status'
                    name: status
                    in: query
            responses:
                '200':
                    description: 'Customer orders retrieved successfully'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message: {type: string}
                                    data: {type: object, properties: {orders: {type: array, items: {type: object, properties: {id: {type: string}, status: {type: string}, paymentStatus: {type: string}, total: {type: number}, createdAt: {type: string, format: date}, updatedAt: {type: [string, 'null'], format: date}, store: {type: object, properties: {id: {type: string}, name: {type: string}}, required: [id, name]}, items: {type: array, items: {type: object, properties: {id: {type: string}, productId: {type: string}, quantity: {type: number}, price: {type: number}, product: {type: object, properties: {id: {type: string}, name: {type: string}, imageUrl: {type: string}}, required: [id, name, imageUrl]}}, required: [id, productId, quantity, price, product]}}}, required: [id, status, paymentStatus, total, createdAt, updatedAt, store, items]}}, pagination: {type: object, properties: {page: {type: number}, limit: {type: number}, total: {type: number}, totalPages: {type: number}}, required: [page, limit, total, totalPages]}}, required: [orders, pagination]}
                                required:
                                    - message
                                    - data
                '400':
                    description: 'Bad request'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message: {type: string}
                                    error: {type: string}
                                required:
                                    - message
                '401':
                    description: Unauthorized
    '/customer/orders/{orderId}':
        get:
            tags:
                - customer
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    schema:
                        type: string
                        description: 'Order ID'
                    required: true
                    description: 'Order ID'
                    name: orderId
                    in: path
            responses:
                '200':
                    description: 'Customer order details retrieved successfully'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message: {type: string}
                                    data: {type: object, properties: {order: {type: object, properties: {id: {type: string}, status: {type: string}, paymentStatus: {type: string}, total: {type: number}, subtotal: {type: number}, tax: {type: number}, shipping: {type: number}, discount: {type: number}, currency: {type: string}, createdAt: {type: string, format: date}, updatedAt: {type: [string, 'null'], format: date}, estimatedDelivery: {type: [string, 'null'], format: date}, deliveredAt: {type: [string, 'null'], format: date}, shippingAddress: {type: [object, 'null'], properties: {street: {type: string}, city: {type: string}, state: {type: string}, postalCode: {type: string}, country: {type: string}}, required: [street, city, state, postalCode, country]}, billingAddress: {type: [object, 'null'], properties: {street: {type: string}, city: {type: string}, state: {type: string}, postalCode: {type: string}, country: {type: string}}, required: [street, city, state, postalCode, country]}, store: {type: object, properties: {id: {type: string}, name: {type: string}, slug: {type: string}}, required: [id, name, slug]}, items: {type: array, items: {type: object, properties: {id: {type: string}, productId: {type: string}, quantity: {type: number}, price: {type: number}, product: {type: object, properties: {id: {type: string}, name: {type: string}, imageUrl: {type: string}}, required: [id, name, imageUrl]}}, required: [id, productId, quantity, price, product]}}}, required: [id, status, paymentStatus, total, subtotal, tax, shipping, discount, currency, createdAt, updatedAt, estimatedDelivery, deliveredAt, shippingAddress, billingAddress, store, items]}}, required: [order]}
                                required:
                                    - message
                                    - data
                '400':
                    description: 'Bad request'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message: {type: string}
                                    error: {type: string}
                                required:
                                    - message
                '401':
                    description: Unauthorized
                '403':
                    description: 'Access denied - customer can only access their own orders'
                '404':
                    description: 'Order not found'
    '/customer/orders/{orderId}/tracking':
        get:
            tags:
                - customer
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    schema:
                        type: string
                        description: 'Order ID'
                    required: true
                    description: 'Order ID'
                    name: orderId
                    in: path
            responses:
                '200':
                    description: 'Order tracking details retrieved successfully'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message: {type: string}
                                    data: {type: object, properties: {orderId: {type: string}, status: {type: string}, paymentStatus: {type: string}, estimatedDelivery: {type: [string, 'null'], format: date}, deliveredAt: {type: [string, 'null'], format: date}, currentLocation: {type: string}, trackingEvents: {type: array, items: {type: object, properties: {eventType: {type: string}, description: {type: string}, timestamp: {type: string, format: date}}, required: [eventType, description, timestamp]}}, deliveryUpdates: {type: array, items: {type: object, properties: {status: {type: string}, location: {type: [string, 'null']}, timestamp: {type: string, format: date}, description: {type: string}}, required: [status, location, timestamp, description]}}}, required: [orderId, status, paymentStatus, estimatedDelivery, deliveredAt, currentLocation, trackingEvents, deliveryUpdates]}
                                required:
                                    - message
                                    - data
                '400':
                    description: 'Bad request'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message: {type: string}
                                    error: {type: string}
                                required:
                                    - message
                '401':
                    description: Unauthorized
                '403':
                    description: 'Access denied - customer can only access their own orders'
                '404':
                    description: 'Order not found'
    /customer/stats:
        get:
            tags:
                - customer
            security:
                -
                    bearerAuth: []
            responses:
                '200':
                    description: 'Customer statistics retrieved successfully'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message: {type: string}
                                    data: {type: object, properties: {totalOrders: {type: number}, totalSpent: {type: number}, averageOrderValue: {type: number}, ordersThisMonth: {type: number}}, required: [totalOrders, totalSpent, averageOrderValue, ordersThisMonth]}
                                required:
                                    - message
                                    - data
                '400':
                    description: 'Bad request'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message: {type: string}
                                    error: {type: string}
                                required:
                                    - message
                '401':
                    description: Unauthorized
    /customer/dashboard:
        get:
            tags:
                - customer
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    schema:
                        type: string
                        enum:
                            - 7d
                            - 30d
                            - 90d
                            - 1y
                            - all
                        default: 30d
                    required: false
                    name: period
                    in: query
            responses:
                '200':
                    description: 'Customer dashboard statistics retrieved successfully'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    totalOrders: {type: number}
                                    totalSpent: {type: number}
                                    averageOrderValue: {type: number}
                                    ordersThisMonth: {type: number}
                                    spendingThisMonth: {type: number}
                                    favoriteStores: {type: array, items: {type: object, properties: {storeId: {type: string}, storeName: {type: string}, orderCount: {type: number}, totalSpent: {type: number}}}}
                                    recentActivity: {type: array, items: {type: object, properties: {eventType: {type: string}, description: {type: string}, timestamp: {type: string}, orderId: {type: string}, orderTotal: {type: number}, storeName: {type: string}}}}
                                    orderStatusBreakdown: {type: object, properties: {pending: {type: number}, processing: {type: number}, shipped: {type: number}, delivered: {type: number}, cancelled: {type: number}}}
                '401':
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message: {type: string}
    /customer/analytics:
        get:
            tags:
                - customer
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    schema:
                        type: number
                        minimum: 1
                        maximum: 365
                        default: 30
                    required: false
                    name: days
                    in: query
            responses:
                '200':
                    description: 'Customer analytics retrieved successfully'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    orderTrends: {type: array, items: {type: object, properties: {date: {type: string}, orderCount: {type: number}}}}
                                    spendingTrends: {type: array, items: {type: object, properties: {date: {type: string}, amount: {type: number}}}}
                                    categoryBreakdown: {type: array, items: {type: object, properties: {productId: {type: string}, quantity: {type: number}}}}
                '401':
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message: {type: string}
    /customer/preferences:
        get:
            tags:
                - customer
            security:
                -
                    bearerAuth: []
            responses:
                '200':
                    description: 'Customer preferences retrieved successfully'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    spendingPatterns: {type: object, properties: {averageOrderValue: {type: number}, preferredOrderDay: {type: string}, preferredOrderTime: {type: string}, totalSpent: {type: number}}}
                                    favoriteProducts: {type: array, items: {type: object, properties: {productId: {type: string}, productName: {type: string}, purchaseCount: {type: number}, lastPurchased: {type: string}}}}
                                    preferredStores: {type: array, items: {type: object, properties: {storeId: {type: string}, storeName: {type: string}, orderCount: {type: number}, totalSpent: {type: number}, averageRating: {type: number}}}}
                                    shoppingBehavior: {type: object, properties: {mostActiveMonth: {type: string}, averageItemsPerOrder: {type: number}, preferredCategories: {type: array, items: {type: string}}, returnRate: {type: number}}}
                '401':
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message: {type: string}
...